emscripten: add command line options to the REPL
