+++ emscripten/main.js
@@ -1,5 +1,5 @@
 
-const mpy = require('./build/firmware.js');
+const mpy = require('./build/micropython.js');
 
 module.exports.run = function(code) {
     mpy.ccall('mp_js_run', 'null', ['string'], [code]);
